move tag generator to tag package
