+++ src/datagen/java/com/github/thedeathlycow/frostiful/datagen/FrostifulDataGenerator.java
@@ -2,7 +2,7 @@ package com.github.thedeathlycow.frostiful.datagen;
 
 import com.github.thedeathlycow.frostiful.Frostiful;
 import com.github.thedeathlycow.frostiful.datagen.generator.FRecipeProvider;
-import com.github.thedeathlycow.frostiful.datagen.generator.ItemTagGenerator;
+import com.github.thedeathlycow.frostiful.datagen.generator.tag.ItemTagGenerator;
 import com.github.thedeathlycow.frostiful.datagen.generator.client.FrostifulModelGenerator;
 import com.github.thedeathlycow.frostiful.registry.FArmorTrimPatterns;
 import com.github.thedeathlycow.frostiful.registry.FBannerPatterns;
